requirements.txt
streamlit
pandas
spacy
python-docx
pdfminer.six
scikit-learn
nltk